Allow user to input conversion rate
Instead of limiting them to options in a selector.

Also, the default is updated to 100, instead of 1, to be more realistic to how most people will be using this. Note that this shouldn't cause anybody's site to suddenly begin using a different value, because if they have enabled the gateway, they will have had to actually save its settings, which will mean the value will be set in the database, not just the default being used.

As a bonus, translator comments are added for the strings using placeholders.

Fixes #16