Gateway: calculate refund from conversion rate of order
In prior versions of the module, we did not save the conversion rate
used in the order. Therefore we had to rely on the current settings to
calculate the amount of points to refund. Now we will have the correct
conversion rate saved for future transaction, but for legacy
transactions we still need to calculate the effective conversion rate
based on the order total and points amount, to avoid refunding the
wrong amount if the conversion rate settings have changed.

Fixes #35